refactor : setter 추가. 메서드 이동
